projects/cardcraft-web/requirements.txt
anyio==4.8.0 ; python_version >= "3.10" and python_version < "4.0"
appnope==0.1.4 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "darwin"
args==0.1.0 ; python_version >= "3.10" and python_version < "4.0"
asgiref==3.8.1 ; python_version >= "3.10" and python_version < "4.0"
asttokens==3.0.0 ; python_version >= "3.10" and python_version < "4.0"
backcall==0.2.0 ; python_version >= "3.10" and python_version < "4.0"
base58==2.1.1 ; python_version >= "3.10" and python_version < "4.0"
blinker==1.9.0 ; python_version >= "3.10" and python_version < "4.0"
boolean-py==4.0 ; python_version >= "3.10" and python_version < "4.0"
cachecontrol[filecache]==0.14.2 ; python_version >= "3.10" and python_version < "4.0"
certifi==2025.1.31 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.4.1 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.10" and python_version < "4.0"
clint==0.5.1 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
construct-typing==0.5.6 ; python_version >= "3.10" and python_version < "4.0"
construct==2.10.68 ; python_version >= "3.10" and python_version < "4.0"
coverage==7.6.12 ; python_version >= "3.10" and python_version < "4.0"
cyclonedx-python-lib==8.9.0 ; python_version >= "3.10" and python_version < "4.0"
decorator==5.2.1 ; python_version >= "3.10" and python_version < "4.0"
defusedxml==0.7.1 ; python_version >= "3.10" and python_version < "4.0"
django==5.1.7 ; python_version >= "3.10" and python_version < "4.0"
dnspython==2.7.0 ; python_version >= "3.10" and python_version < "4.0"
exceptiongroup==1.2.2 ; python_version >= "3.10" and python_version < "3.11"
executing==2.2.0 ; python_version >= "3.10" and python_version < "4.0"
expects==0.9.0 ; python_version >= "3.10" and python_version < "4.0"
filelock==3.17.0 ; python_version >= "3.10" and python_version < "4.0"
flask[async]==3.0.3 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
httpcore==1.0.7 ; python_version >= "3.10" and python_version < "4.0"
httptools==0.6.4 ; python_version >= "3.10" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.10" and python_version < "4.0"
idna==3.10 ; python_version >= "3.10" and python_version < "4.0"
ipython==8.10.0 ; python_version >= "3.10" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.10" and python_version < "4.0"
jedi==0.19.2 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.6 ; python_version >= "3.10" and python_version < "4.0"
jsonalias==0.1.1 ; python_version >= "3.10" and python_version < "4.0"
license-expression==30.4.1 ; python_version >= "3.10" and python_version < "4.0"
mamba==0.11.3 ; python_version >= "3.10" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.10" and python_version < "4.0"
matplotlib-inline==0.1.7 ; python_version >= "3.10" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.10" and python_version < "4.0"
motor==3.6.0 ; python_version >= "3.10" and python_version < "4.0"
msgpack==1.1.0 ; python_version >= "3.10" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
mypy==1.11.2 ; python_version >= "3.10" and python_version < "4.0"
packageurl-python==0.16.0 ; python_version >= "3.10" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.10" and python_version < "4.0"
parso==0.8.4 ; python_version >= "3.10" and python_version < "4.0"
pexpect==4.9.0 ; python_version >= "3.10" and python_version < "4.0" and sys_platform != "win32"
pickleshare==0.7.5 ; python_version >= "3.10" and python_version < "4.0"
pip-api==0.0.34 ; python_version >= "3.10" and python_version < "4.0"
pip-audit==2.8.0 ; python_version >= "3.10" and python_version < "4.0"
pip-requirements-parser==32.0.1 ; python_version >= "3.10" and python_version < "4.0"
pip==25.0.1 ; python_version >= "3.10" and python_version < "4.0"
platformdirs==4.3.6 ; python_version >= "3.10" and python_version < "4.0"
prompt-toolkit==3.0.50 ; python_version >= "3.10" and python_version < "4.0"
ptyprocess==0.7.0 ; python_version >= "3.10" and python_version < "4.0" and sys_platform != "win32"
pure-eval==0.2.3 ; python_version >= "3.10" and python_version < "4.0"
py-serializable==1.1.2 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.10" and python_version < "4.0"
pygments==2.19.1 ; python_version >= "3.10" and python_version < "4.0"
pyhiccup==0.1 ; python_version >= "3.10" and python_version < "4.0"
pymongo==4.9.2 ; python_version >= "3.10" and python_version < "4.0"
pynacl==1.5.0 ; python_version >= "3.10" and python_version < "4.0"
pyparsing==3.2.1 ; python_version >= "3.10" and python_version < "4.0"
pyrsistent==0.20.0 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==0.20.0 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.10" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.10" and python_version < "4.0"
rich==13.9.4 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
solana==0.34.3 ; python_version >= "3.10" and python_version < "4.0"
solders==0.21.0 ; python_version >= "3.10" and python_version < "4.0"
sortedcontainers==2.4.0 ; python_version >= "3.10" and python_version < "4.0"
sqlparse==0.5.3 ; python_version >= "3.10" and python_version < "4.0"
stack-data==0.6.3 ; python_version >= "3.10" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.10" and python_version < "4.0"
tomli==2.2.1 ; python_version >= "3.10" and python_version < "3.11"
traitlets==5.14.3 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2025.1 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
urllib3==2.3.0 ; python_version >= "3.10" and python_version < "4.0"
uvicorn[standard]==0.30.6 ; python_version >= "3.10" and python_version < "4.0"
uvloop==0.21.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.10" and python_version < "4.0"
watchfiles==1.0.4 ; python_version >= "3.10" and python_version < "4.0"
wcwidth==0.2.13 ; python_version >= "3.10" and python_version < "4.0"
websockets==11.0.3 ; python_version >= "3.10" and python_version < "4.0"
werkzeug==3.0.6 ; python_version >= "3.10" and python_version < "4.0"
pyinstaller==6.12.0
